+++ tools/clean_swp_files.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/bash
+find . -name "*.swp" -delete
+find . -name "*.swo" -delete
+find . -name "*.swn" -delete
